Split Buildvana.Tool.Cli into Commands, CommandLine, and Infrastructure.Execution
Unclutter the overloaded Cli namespace by extracting cohesive groups:

- Buildvana.Tool.Commands: the command classes plus their settings
  (BaseSettings, ReleaseSettings, GlobalOptions) and BuildSteps.
- Buildvana.Tool.CommandLine: ForwardedArguments (the parsing primitives
  follow in the next commit).
- Buildvana.Tool.Infrastructure.Execution: command discovery and dispatch
  (ImplementsCommandAttribute, CommandRegistration, CommandRegistry).

The Spectre.Console.Cli glue (BvHelpProvider, TypeRegistrar, TypeResolver)
stays in Cli pending its removal. Behavior-preserving: Spectre still drives
parsing and dispatch. GitService aliases LibGit2Sharp.Commands so the new
Commands namespace no longer shadows it.
