+++ src/Buildvana.Tool/Cli/BvHelpProvider.cs
@@ -5,6 +5,8 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using System.Reflection;
+using Buildvana.Tool.Commands;
+using Buildvana.Tool.Infrastructure.Execution;
 using Spectre.Console;
 using Spectre.Console.Cli;
 using Spectre.Console.Cli.Help;
