+++ src/Buildvana.Tool/Commands/GlobalOptions.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (C) Tenacom and Contributors. Licensed under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-namespace Buildvana.Tool.Cli;
+namespace Buildvana.Tool.Commands;
 
 /// <summary>
 /// Bv-global options pre-parsed from the command-line args in <c>Program.Main</c> before Spectre dispatches the subcommand.
