+++ src/Buildvana.Tool/Commands/RestoreCommand.cs
@@ -5,19 +5,21 @@ using System;
 using System.ComponentModel;
 using System.Threading;
 using System.Threading.Tasks;
+using Buildvana.Tool.Infrastructure.Execution;
 using CommunityToolkit.Diagnostics;
 using Spectre.Console.Cli;
 
-namespace Buildvana.Tool.Cli;
+namespace Buildvana.Tool.Commands;
 
-[ImplementsCommand("clean")]
-[Description("Remove all build artifacts, intermediate output, and temporary files. Like 'dotnet clean', but more aggressive.")]
-internal sealed class CleanCommand(IServiceProvider services) : AsyncCommand<BaseSettings>
+[ImplementsCommand("restore", consumesAllArguments: true)]
+[Description("Clean and restore dependencies.")]
+internal sealed class RestoreCommand(IServiceProvider services) : AsyncCommand<BaseSettings>
 {
     protected override async Task<int> ExecuteAsync(CommandContext context, BaseSettings settings, CancellationToken cancellationToken)
     {
         Guard.IsNotNull(settings);
         await BuildSteps.CleanAsync(services).ConfigureAwait(false);
+        await BuildSteps.RestoreAsync(services).ConfigureAwait(false);
         return 0;
     }
 }
