+++ src/Buildvana.Tool/Program.cs
@@ -12,7 +12,10 @@ using Buildvana.Core.HomeDirectory;
 using Buildvana.Core.Json;
 using Buildvana.Core.Process;
 using Buildvana.Tool.Cli;
+using Buildvana.Tool.CommandLine;
+using Buildvana.Tool.Commands;
 using Buildvana.Tool.Configuration;
+using Buildvana.Tool.Infrastructure.Execution;
 using Buildvana.Tool.Infrastructure.Logging;
 using Buildvana.Tool.Services;
 using Buildvana.Tool.Services.Git;
