+++ src/Buildvana.Tool/Services/DotNetService.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Buildvana.Tool.Cli;
+using Buildvana.Tool.Commands;
 using Buildvana.Tool.Configuration;
 using Buildvana.Tool.Infrastructure;
 using Buildvana.Tool.Services.ServerAdapters;
